Change Builder structure
If a Builder was appended to an existing Builder before, it was
theoretically supposed to move the strings to the existing instance.
That wasn't what happened, and the Builder was just added to form a tree of
Builders and Strings. In hindsight, that makes more sense for a string
builder, so that's what it does intentionally now.